refactor(web): modularize bootstrap and split shell e2e scenarios
