+++ apps/web/src/main/feature-contract.ts
@@ -1,0 +1,5 @@
+export interface FeatureModule<TActions> {
+  init(): void;
+  dispose(): void;
+  actions: TActions;
+}
