+++ apps/web/test/e2e/web-shell-tool-flow/scenarios/chat-list-search.scenario.ts
@@ -1,0 +1,5 @@
+// Aggregate file: keep stable import path, tests are split into smaller scenario files.
+
+import "./chat-list-search/live-refresh.scenario";
+import "./chat-list-search/search-and-qq.scenario";
+import "./chat-list-search/delete-and-switch.scenario";
