+++ apps/web/test/e2e/web-shell-tool-flow/scenarios/cron.scenario.ts
@@ -1,0 +1,5 @@
+// Aggregate file: keep stable import path, tests are split into smaller scenario files.
+
+import "./cron/text-job.scenario";
+import "./cron/workflow.scenario";
+import "./cron/panel-and-protection.scenario";
