+++ apps/web/test/e2e/web-shell-tool-flow/scenarios/prompt-mode.scenario.ts
@@ -1,0 +1,5 @@
+// Aggregate file: keep stable import path, tests are split into smaller scenario files.
+
+import "./prompt-mode/template-expansion.scenario";
+import "./prompt-mode/context-introspect.scenario";
+import "./prompt-mode/prompt-mode-routing.scenario";
