+++ apps/web/test/e2e/web-shell-tool-flow/scenarios/shell-rendering.scenario.ts
@@ -1,0 +1,5 @@
+// Aggregate file: keep stable import path, tests are split into smaller scenario files.
+
+import "./shell-rendering/command-and-stream.scenario";
+import "./shell-rendering/tool-call-summary.scenario";
+import "./shell-rendering/history-and-merge.scenario";
