+++ apps/web/test/e2e/web-shell-tool-flow/scenarios/workspace-settings.scenario.ts
@@ -1,0 +1,5 @@
+// Aggregate file: keep stable import path, tests are split into smaller scenario files.
+
+import "./workspace-settings/qq-channel.scenario";
+import "./workspace-settings/editor-modal.scenario";
+import "./workspace-settings/cards.scenario";
