+++ apps/web/test/e2e/web-shell-tool-flow/support/test-helpers.ts
@@ -1,0 +1,48 @@
+import { readFileSync } from "node:fs";
+import { join } from "node:path";
+
+import { afterEach, beforeEach, vi } from "vitest";
+
+export function jsonResponse(body: unknown, status = 200): Response {
+  return new Response(JSON.stringify(body), {
+    status,
+    headers: {
+      "content-type": "application/json",
+    },
+  });
+}
+
+export async function waitFor(condition: () => boolean, timeoutMS = 2000): Promise<void> {
+  const startedAt = Date.now();
+  while (!condition()) {
+    if (Date.now() - startedAt > timeoutMS) {
+      throw new Error("timeout waiting for condition");
+    }
+    await new Promise((resolve) => setTimeout(resolve, 10));
+  }
+}
+
+export async function mountWebApp(): Promise<void> {
+  await import("../../../../src/main.ts");
+}
+
+export function useWebShellFlowFixture(): void {
+  let originalFetch: typeof globalThis.fetch;
+
+  beforeEach(() => {
+    vi.resetModules();
+    document.documentElement.innerHTML = readFileSync(join(process.cwd(), "src/index.html"), "utf8").replace(
+      /<!doctype html>/i,
+      "",
+    );
+    window.localStorage.clear();
+    window.localStorage.setItem("nextai.web.locale", "zh-CN");
+    originalFetch = globalThis.fetch;
+  });
+
+  afterEach(() => {
+    vi.restoreAllMocks();
+    globalThis.fetch = originalFetch;
+    document.documentElement.innerHTML = "<head></head><body></body>";
+  });
+}
